Fix library materials and schedule board editing

+++ public/js/library/EducationMaterials.js
@@ -1,14 +1,25 @@
 
 $( document ).ready(function() {
 //Удаление материала
-    $('.deleteEducationalMaterial').click(function () {
+    $(document)
+        .off('click.educationalMaterials', '.deleteEducationalMaterial')
+        .on('click.educationalMaterials', '.deleteEducationalMaterial', function (event) {
+        event.preventDefault();
+        event.stopImmediatePropagation();
+
+        var button = $(this);
+        if (button.data('deleting')) {
+            return false;
+        }
+
         var x = confirm("Удалить материал?");
         if (x) {
-            var id = $(this).attr("id");
-            var token = $(this).attr("value");
+            button.data('deleting', true);
+            var id = button.attr("id");
+            var token = button.attr("value");
             $.ajax(
                 {
-                    url: "educationalMaterials/" + id + "/delete",
+                    url: "/library/educationalMaterials/" + id + "/delete",
                     type: 'DELETE',
                     dataType: "JSON",
                     data: {
@@ -23,6 +34,14 @@ $( document ).ready(function() {
 
                             $('[name = "delete'+data.id+'"]').parent().parent().hide();
                         }
+                    },
+                    error: function (xhr) {
+                        button.data('deleting', false);
+                        var message = 'Не удалось удалить материал.';
+                        if (xhr.responseJSON && xhr.responseJSON.message) {
+                            message += ' ' + xhr.responseJSON.message;
+                        }
+                        alert(message);
                     }
                 });
         }
